backend/requirements.txt
django==4.2.6
djangorestframework==3.14.0
redis==4.6.0
psycopg2-binary==2.9.6
pillow==10.0.0
django-cors-headers==4.2.0
cryptography==3.4.8
django-cors-headers==4.2.0
django-sslserver==0.22
channels==4.0.0
daphne==4.0.0
channels_redis==4.1.0
watchdog==3.0.0
asgiref==3.7.2
whitenoise==6.5.0
django-ipware==5.0.0
requests==2.28.2
Celery==5.3.4
flower==2.0.1
eventlet==0.33.3
gunicorn==21.2.0
uvicorn[standard]==0.23.2
inotify==0.2.10
steam[client]==1.4.4
python-decouple==3.8